% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread-data.R
\docType{data}
\name{bread}
\alias{bread}
\title{Bread data}
\format{
An object of class \code{"dataframe"}.
}
\usage{
data(bread)
}
\description{
Data from a sensory experiment
with data on 315 likert responses about appearance,
aroma, taste, texture, color of bread.
}
\examples{
data(grav)
samplen(bread$color, 10)
}
\references{
FQAP2000798
}
\keyword{datasets}
