% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancake-data.R
\docType{data}
\name{pancake}
\alias{pancake}
\title{pancake data}
\format{
An object of class \code{"dataframe"}.
}
\usage{
data(pancake)
}
\description{
Data from a sensory experiment
with data on 315 likert responses about acceptance, appearance,
aroma, taste, texture, and color of pancake.
}
\examples{
data(pancake)
samplen(pancake$color, 10)
}
\references{
FQAP2000798
}
\keyword{datasets}
