% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplen.R
\name{samplen}
\alias{samplen}
\title{sample size by boostrap}
\usage{
samplen(data, n, typeCI = 3, nboot = 1000)
}
\arguments{
\item{data}{A numerical vector.}

\item{n}{An integer.}
}
\value{
An confidence interval for the error mean.
}
\description{
Estimate the bootstrap error
}
\examples{
data(cookie)
samplen(cookie$color, 31)
# male
samplen(cookie$color[cookie$gender==1], 35)
}
