% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfiha-data.R
\docType{data}
\name{sfiha}
\alias{sfiha}
\title{Sfiha data}
\format{
An object of class \code{"dataframe"}.
}
\usage{
data(sfiha)
}
\description{
Data from a sensory experiment
with data on 315 likert responses about appearance,
aroma, taste, texture, color of sfiha.
}
\examples{
data(grav)
samplen(sfiha$color, 10)
}
\references{
FQAP2000798
}
\keyword{datasets}
