% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie-data.R
\docType{data}
\name{cookie}
\alias{cookie}
\title{Cookie data}
\format{
An object of class \code{"dataframe"}.
}
\usage{
data(cookie)
}
\description{
Data from a sensory experiment
with data on 315 likert responses about appearance,
aroma, taste, texture, color of cookie.
}
\examples{
data(grav)
samplen(cookie$color, 10)
}
\references{
FQAP2000798
}
\keyword{datasets}
