% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muffin-data.R
\docType{data}
\name{muffin}
\alias{muffin}
\title{Muffin data}
\format{
An object of class \code{"dataframe"}.
}
\usage{
data(muffin)
}
\description{
Data from a sensory experiment
with data on 315 likert responses about acceptance, appearance,
aroma, taste, texture, and color of muffin.
}
\examples{
data(muffin)
samplen(muffin$color, 10)
}
\references{
FQAP2000798
}
\keyword{datasets}
